Fix test usage of user.Name when user has full name set
When a Linux user has a full name set, the Go user.Name field returns
the full name, rather than the username. The audit_test uses this
value for the output.file.user config, which causes the test to fail
with an unexpected "Could not find uid for user" error returned by
createFileOutput(). This patch fixes the test to use the Username field
instead of the Name field.