Add GELF output to be able to send logs to a Graylog2 server
This change adds a new output option: gelf. It can be used to send messages
with GELF standard to a compliant Graylog2 server. GELF provides several
features that a simple syslog message does not as chunked and compressed
messages, e.g., see more details in [1].

Most the settings of GELF can be handled on go-audit configuration file.

[1] http://docs.graylog.org/en/stable/pages/gelf.html